Deprecated API:

StableHashTraits.hash_method — Function

StableHashTraits.hash_method(x, [context])

Retrieve the trait object that indicates how a type should be hashed using stable_hash. You should return one of the following values.

WriteHash(): writes the object to a binary format using StableHashTraits.write(io, x) and takes a hash of that. StableHashTraits.write(io, x) falls back to Base.write(io, x) if no specialized methods are defined for x.

IterateHash(): assumes the object is iterable and finds a hash of all elements

StructHash([pair = (fieldnames ∘ typeof) => getfield], [order]): hash the structure of the object as defined by a sequence of pairs. How precisely this occurs is determined by the two arguments:

pair Defines how fields are extracted; the default is fieldnames ∘ typeof => getfield but this could be changed to e.g. propertynames => getproperty or Tables.columnnames => Tables.getcolumn. The first element of the pair is a function used to compute a list of keys and the second element is a two argument function used to extract the keys from the object.

order can be :ByOrder (the default)—which sorts by the order returned by pair[1]—or :ByName—which sorts by lexigraphical order.

FnHash(fn, [method]): hash the result of applying fn to the given object. Optionally, use method to hash the result of fn, otherwise calls hash_method on the result to determine how to hash it. There are two built-in functions commonly used with FnHash.

stable_typename_id: Get the qualified name of an object's type, e.g. Base.String and return 64 bit hash of this string

stable_type_id: Get the qualified name and type parameters of a type, e.g. Base.Vector{Int}, and return a 64 bit hash of this string.

@ConstantHash(x): at compile time, hash the literal (constant) string or number using sha256 and include the first 64 bits as a constant number that is recursively hashed using the WriteHash method.

Tuple: apply multiple methods to hash the object, and then recursively hash their results. For example: (@ConstantHash("header"), StructHash()) would compute a hash for both the string "header" and the fields of the object, and then recursively hash these two hashes.

Your hash will be stable if the output for the given method remains the same: e.g. if write is the same for an object that uses WriteHash, its hash will be the same; if the fields are the same for StructHash, the hash will be the same; etc...

Missing from the above list is one final, advanced, trait: HashAndContext which can be used to change the context within the scope of a given object.

Customizing hash computations with contexts

You can customize how hashes are computed within a given scope using a context object. This is also a very useful way to avoid type piracy. The context can be any object you'd like and is passed as the second argument to stable_hash. By default it is equal to HashVersion{1}() and this determines how objects are hashed when a more specific method is not defined.

This context is then passed to both hash_method and StableHashTraits.write (the latter is the method called for WriteHash, and falls back to Base.write). Because of the way the root contexts (HashVersion{1} and HashVersion{2}) are defined, you normally don't have to include this context as an argument when you define a method of hash_context or write because there are appropriate fallback methods.

When you define a hash context it should normally accept a parent context that serves as a fallback, and return it in an implementation of the method StableHashTraits.parent_context.

As an example, here is how we could write a context that treats all named tuples with the same keys and values as equivalent.

struct NamedTuplesEq{T}
    parent::T
end
StableHashTraits.parent_context(x::NamedTuplesEq) = x.parent
function StableHashTraits.hash_method(::NamedTuple, ::NamedTuplesEq)
    return FnHash(stable_typename_id), StructHash(:ByName)
end
context = NamedTuplesEq(HashVersion{2}())
stable_hash((; a=1:2, b=1:2), context) == stable_hash((; b=1:2, a=1:2), context) # true

If we instead defined parent_context to return nothing, our context would need to implement a hash_method that covered the types AbstractRange, Int64, Symbol and Pair for the call to stable_hash above to succeed.

Customizing hashes within an object

Contexts can be customized not only when you call stable_hash but also when you hash the contents of a particular object. This lets you change how hashing occurs within the object. See the docstring of HashAndContext for details.

StableHashTraits.stable_typename_id — Function

stable_typename_id(x)

Returns a 64 bit hash that is the same for a given type so long as the name and the module of the type doesn't change.

Example

julia> stable_typename_id([1, 2, 3])
0x56c6b9ca080a0aa4

julia> stable_typename_id(["a", "b"])
0x56c6b9ca080a0aa4

Note

If the module of a type is Core it is renamed to Base before hashing because the location of some types changes between Core to Base across julia versions. Likewise, the type names of AbstractArray types are made uniform as their printing changes from Julia 1.6 -> 1.7.

StableHashTraits.stable_type_id — Function

stable_type_id(x)

Returns a 64 bit hash that is the same for a given type so long as the module, and string representation of a type is the same (invariant to comma spacing).

Example

julia> stable_type_id([1, 2, 3])
0xfd5878e59e259648

julia> stable_type_id(["a", "b"])
0xe191f67c4c8e3370

Note

If the module of a type is Core it is renamed to Base before hashing because the location of some types changes between Core to Base across julia versions. Likewise, the type names of AbstractArray types are made uniform as their printing changes from Julia 1.6 -> 1.7.

Warn

This function has a known bug that has been left in to avoid breaking old hashes. (The long-term plan is to eliminate this and qualified_type from the API, see https://github.com/beacon-biosignals/StableHashTraits.jl/pull/55 for details). The bug means that the type has can depend on the order in which you load modules and call stable_type_id. To make use of the version of this function that has been fixed you can call stable_type_id_fixed

StableHashTraits.write — Function

StableHashTraits.write(io, x, [context])

Writes contents of x to an io buffer to be hashed during a call to stable_hash. Fall back methods are defined as follows:

write(io, x, context) = write(io, x)
write(io, x) = Base.write(io, x)

Users of StableHashTraits can overwrite either the 2 or 3 argument version for their types to customize the behavior of stable_hash.

See also: StableHashTraits.hash_method.

StableHashTraits.HashAndContext — Type

HashAndContext(method, old_context -> new_context)

A special hash method that changes the context when hashing the contents of an object. The method defines how the object itself should be hashed and the second argument is a callable which transforms the old context to the new.

For example, here is how we can make sure the arrays in a specific object have a hash that is invariant to endianness without having to copy the array.

struct EndianInvariant{P}
    parent_context::P
end
StableHashTraits.parent_context(x::EndianInvariant) = x.parent_context

struct CrossPlatformData
    data::Vector{Int}
end

StableHashTraits.hash_method(::Number, ::EndianInvariant) = FnHash(htol, WriteHash())
StableHashTraits.hash_method(::CrossPlatformData) = HashAndContext(IterateHash(), EndianInvariant)

Note that we could accomplish this same behavior using FnHash(x -> htol.(x.data)), but it would require copying that data to do so.