Internal Methods

Documented methods that are not part of the public API.

StableHashTraits.hash_type! — Function

hash_type!(hash_state, context, T)

Hash type T in the given context, updating hash_state.

StableHashTraits.HashState — Type

HashState(alg, context)

Given a function that implements the hash algorithm to use and the current hash context, setup the necessary state to track updates to hashing as we traverse an object's structure and return it.

StableHashTraits.similar_hash_state — Function

similar_hash_state(state::HashState)

Akin to similar for arrays, this constructs a new object of the same concrete type as state

StableHashTraits.start_nested_hash! — Function

start_nested_hash!(state::HashState)

Return an updated state that delimits hashing of a nested structure; calls made to update_hash! after startnestedhash! will be handled as nested elements up until end_nested_hash! is called.

StableHashTraits.end_nested_hash! — Function

end_nested_hash!(state::HashState)

Return an updated state that delimints the end of a nested structure.

StableHashTraits.update_hash! — Function

update_hash!(state::HashState, bytes)

Returns the updated hash state given a set of bytes (either a tuple or array of UInt8 values).

update_hash!(state::HashState, obj, context)

Returns the updated hash, given an object and some context. The object will be written to some bytes using StableHashTraits.write(io, obj, context).

StableHashTraits.compute_hash! — Function

compute_hash!(state::HashState)

Return the final hash value to return for state

Deprecated, non-public API methods; these will be deleted when the deprecated functionality is removed.

StableHashTraits.stable_typefields_id — Function

stable_typefields_id(x)

Returns a 64 bit hash that is the same for a given type so long as the set of field names remains unchanged.