API

StableHashTraits.stable_hash — Function

stable_hash(x, context=HashVersion{1}(); alg=sha256)
stable_hash(x; alg=sha256, version=1)

Create a stable hash of the given objects. As long as the context remains the same, this hash is intended to remain unchanged across julia versions. The built-in context is HashVersion{N}, and if you specify a version, this is equivalent to explicitly passing HashVersion{version}. To customize how the hash is copmuted see Using Contexts.

It is best to pass an explicit version, since HashVersion{3} is the only non-deprecated version; it is much faster than 1 and more stable than 2. Furthermore, a new hash version is provided in a future release, the hash you get by passing an explicit HashVersion{N} should not change. (Note that the number in HashVersion does not necessarily match the package version of StableHashTraits).

In hash version 3, you customize how hashes are computed using transformer, and in versions 1-2 using hash_method.

To change the hash algorithm used, pass a different function to alg. It accepts any sha related function from SHA or any function of the form hash(x::AbstractArray{UInt8}, [old_hash]).

The context value gets passed as the second argument to hash_method and [transformer], and as the third argument to StableHashTraits.write. Note that both hash_method and StableHashTraits.write are deprecated.

See Also

transformer

StableHashTraits.HashVersion — Type

HashVersion{V}()

The default hash_context used by stable_hash. There are currently three versions (1-3). Version 3 should be favored when at all possible. Version 1 is the default version, so as to avoid changing the hash computed by existing code.

By explicitly passing this hash version in stable_hash you ensure that hash values for these fallback methods will not change even if new hash versions are developed.

StableHashTraits.stable_type_name — Function

stable_type_name(::Type{T})
stable_type_name(T::Module)
stable_type_name(::T) where {T}

Get a stable name of T. The stable name includes the name of the module that T was defined in. Any uses of Core are replaced with Base to keep the name stable across versions of julia. Anonymous names (e.g. stable_type_name(x -> x+1)) throw an error, as no stable name is possible in this case.

StableHashTraits.transformer — Function

StableHashTraits.transformer(::Type{T}, [context]) where {T}

Return Transformer indicating how to modify an object of type T before hashing it. Methods without a context are called if no method for that type exists for any specific context object.

StableHashTraits.Transformer — Type

StableHashTraits.Transformer(fn=identity, result_method=nothing;
                             preserves_structure=StableHashTraits.preserves_structure(fn))

Wraps the function used to transform values before they are hashed. The function is applied (fn(x)), and then its result is hashed according to the trait @something result_method StructType(fn(x)).

The flag preserves_structure indicates if it is safe to hoist type hashes outside of loops; this is always the case when fn is type stable. See the manual for details about other cases when it is safe to set this flag to true.

See Also

transformer

StableHashTraits.TransformIdentity — Type

StableHashTraits.TransformIdentity(x)

Signal that the type x should not be transformed in the usual way, but by hashing x directly. This is useful when you want to hash both x the way it would normally be hashed without a specialized method of transformer along with some metadata. Without this wrapper, returning (metadata(x), x) from the transforming function would cause an infinite regress (adding metadata(x) upon each call).

Example

struct MyArray <: AbstractVector{Int}
    data::Vector{Int}
    meta::Dict{String, String}
end
# other array methods go here...
function StableHashTraits.transformer(::Type{<:MyArray})
    return Transformer(x -> (x.meta, TransformIdentity(x)); preserves_structure=true)
end

In this example we hash both some metadata about a custom array, and each of the elements of x

StableHashTraits.preserves_structure — Function

StableHashTraits.preserves_structure(fn)

Returns true if it is known that fn preservess structure ala Transformer. This is false by default for all functions but identity. You can define a method of this function for your own fn's to signal that they their results can be safely optimized via hoisting the type hash outside of loops during hashing.

StableHashTraits.CachedHash — Type

CachedHash(context)

Setup a hash context that includes a cache of hash results. By default, it stores the result of hashing types and large values. Calling the same cached hash context will re-use this cache, possibly improving performance. If you do not pass a CachedHash to stable_hash it sets up its own internal cache to improve performance for repeated hashes of the same type or large values seen within the call to stable_hash.

For an object to be cached you must either signal that it should be, using UseCache or it must be:

large enough: this is to ensure that calls to retrieve a cached result do not exceed the time it takes to simply re-hash the individual bytes of an object. You can refer to the constant CACHE_OBJECT_THRESHOLD though it is not considered part of the public API and may change with future hash versions. (The threshold will not change for a given hash version, since changing it can change an object's hashed value).

mutable: immutable objects cannot be stored in a WeakKeyIdDict, since they aren't supported. This means that caching immutable objects can lead to memory leaks if you don't clean up the cache regularly, since they are stored in an IdDict. Note that in practice large amounts of data are usually stored in mutable structures like Array and String.

See Also

stable_hash

StableHashTraits.UseCache — Type

StableHashTraits.UseCache(x)

Signal that the hash of x should be stored in the cache.

Immutable objects can leak memory

If x is immutable, caching it will cause the object x to be held in memory until the cache is garbage collected. WeakKeyIdDicts do not support immutable objects. If there is no user defined cache, the cache will be garbage collected inside the call to stable_hash. With a user defined hash you will need to make sure your cache goes out of scope in a timely fashion to avoid memory leaks.

See Also

transformer

StableHashTraits.type_hash_name — Function

type_hash_name(::Type{T}, trait, [context])

The name that is hashed for type T when hashing the type of a given value. This defaults to stable_type_name(trait). Users of StableHashTraits can implement a method that accepts two (T and trait) or three arguments (T, trait, context). The trait is one of the StructTypes traits and context is the hash context (as per the second argument to stable_hash)

StableHashTraits.type_value_name — Function

type_value_name(::Type{T}, trait, [context]) where {T}

The name that is hashed for type T when hashing a type as a value (e.g. stable_hash(Int)). This defaults to stable_type_name(T). Users of StableHashTraits can implement a method that accepts two (T and trait) or three arguments (T, trait, context). The trait is one of the StructTypes traits and context is the hash context (as per the second argument to stable_hash)

StableHashTraits.type_structure — Function

type_structure(::Type{T}, trait, [context])

Get the types and symbols that represent the structure of T. The trait is the StructType of T and context is the hash context (as per the second argument to stable_hash). The second argument must be included as part of a method definition, but context is optional.

You need to override this method when a type parameter of T is important to the type's hash but is not included as part of the normal hash.

The normal hash includes:

fieldtypes(T) of any StructType.DataType (e.g. StructType.Struct)

eltype(T) of any StructType.ArrayType or StructType.DictType

ndims(T) of any AbstractArray that is a StructType.ArrayTYpe and that has a concrete dimension

Example

struct MyType{T,F}
    obj::T
end
StableHashTratis.type_structure(::Type{T}, ::StructTypes.DataType) where {F,T<:MyType{<:Any,F}}
    return fieldnames(T), fieldtypes(T), F
end

Without this definition MyType{Int,:foo}(1) would hash to the same value as MyType{Int,:bar}(1).

StableHashTraits.is_ordered — Function

is_ordered(x)

Indicates whether the order of the elements of object x are important to its hashed value. If false, x's elements will first be collected and sort'ed before hashing them. When calling sort, hash_sort_by is passed as the by keyword argument. If x is a DictType, the elements are sorted by their keys rather than their elements.

StableHashTraits.hash_sort_by — Function

`hash_sort_by(x)`

Defines how the elements of a hashed container x are sorted if is_ordered of x returns false. The return value of this function is passed to sort as the by keyword.

StableHashTraits.parent_context — Function

StableHashTraits.parent_context(context)

Return the parent context of the given context object. (See hash_method for details of using context). The default method falls back to returning HashVersion{1}, but this is flagged as a deprecation warning; in the future it is expected that all contexts define this method.

This is normally all that you need to know to implement a new context. However, if your context is expected to be the root context—one that does not fallback to any parent (akin to HashVersion)—then there may be a bit more work involved. In this case, parent_context should return nothing. You will also need to define StableHashTraits.root_version.

StableHashTraits.root_version — Function

StableHashTraits.root_version(context)

Return the version of the root context: an integer in the range (1, 2). The default fallback method value returns 1.

In almost all cases, a root hash context should return 2. The optimizations used in HashVersion{2} include a number of changes to the hash-trait implementations that do not alter the documented behavior but do change the actual hash value returned because of how and when elements get hashed.

StableHashTraits.TablesEq — Type

TablesEq(parent_context)

In this hash context the type and structure of a table do not impact the hash that is created, only the set of columns (as determined by Tables.columns), and the hash of the individual columns matter.

StableHashTraits.ViewsEq — Type

ViewsEq(parent_context)

Create a hash context where only the contents of an array or string determine its hash: that is, the type of the array or string (e.g. SubString vs. String) does not impact the hash value.

Deprecated

In HashVersion{3} this is already true, so there is no need for ViewsEq. This does not change the behavior of HashVersion{3} or later.

StableHashTraits.WithTypeNames — Type

WithTypeNames(parent_context)

In this hash context, not only the structure, but also the name of the type (e.g. Array vs. SubArray) affects the hashed value.